requirements.txt
numpy
scipy
matplotlib
cv2
pillow
sklearn
shapely
pandas
geopandas
networkx
osmnx
potrace
pytorch
